Move sortable list to own component
